+++ src/main/resources/templates/layout/dashboard-navibar.html
@@ -11,7 +11,7 @@
                     <div th:fragment="homeNav">
                         <ul class="nav navbar-nav ml-auto">
                             <li class="nav-item active">
-                                <a class="nav-link" href="#">Home</a>
+                                <a class="nav-link" th:href="@{/}">Home</a>
                             </li>
                             <li class="nav-item">
                                 <a class="nav-link" href="#">Corporate</a>
